refactor: split body validators into scene_checks_body
scene_checks.py had grown past the 1000-line module cap with the
body-appearance validators (pose scatter, relief/photometry, albedo
and disc texture, transits).  Those checkers and their key inventories
move to the sibling scene_checks_body module, which builds on the same
scalar primitives; scene.py imports _check_body_object from its new
home.  No behavior change.
